onnx_web/requirements.txt
ftfy
regex
onnxruntime==1.15.0
faiss-cpu==1.7.2
fastapi==0.79.1
lmdb==1.3.0
numpy==1.23.2
Pillow==9.2.0
pydantic==1.9.2
tqdm==4.64.0
uvicorn==0.18.2
python-multipart